Split up RestIT test into separate ( happy flow test ) and endPoint test class, place shared functions in TestUtils
