+++ molgenis-data-rest/src/test/java/org/molgenis/data/rest/RestControllerIT.java
@@ -1,61 +1,36 @@
 package org.molgenis.data.rest;
 
-import com.google.common.collect.ImmutableMap;
-import com.google.common.io.Resources;
 import io.restassured.RestAssured;
-import io.restassured.http.ContentType;
 import io.restassured.response.ValidatableResponse;
-import net.minidev.json.JSONObject;
 import org.elasticsearch.common.Strings;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.mock.web.MockHttpServletResponse;
 import org.testng.annotations.AfterClass;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
 
-import java.io.File;
-import java.net.URI;
-import java.net.URISyntaxException;
-import java.net.URL;
-import java.util.Map;
-
-import static com.google.api.client.util.Maps.newHashMap;
-import static com.google.common.collect.ImmutableMap.of;
-import static com.google.common.collect.Lists.newArrayList;
-import static io.restassured.RestAssured.*;
-import static io.restassured.config.EncoderConfig.encoderConfig;
-import static java.util.Collections.singletonList;
+import static io.restassured.RestAssured.baseURI;
+import static io.restassured.RestAssured.given;
 import static org.hamcrest.Matchers.equalTo;
 import static org.molgenis.data.rest.convert.RestTestUtils.*;
 
 public class RestControllerIT
 {
 	private static final Logger LOG = LoggerFactory.getLogger(RestControllerIT.class);
+	private static final String SYS_SEC_USER_AUTHORITY_ID = "sys_sec_UserAuthority_ID";
 
-	// Request parameters
-	private static final String X_MOLGENIS_TOKEN = "x-molgenis-token";
-	private static final String TEXT_PLAIN = "text/plain";
-	private static final String APPLICATION_FORM_URL_ENCODED = "application/x-www-form-urlencoded; charset=UTF-8";
-	private static final String TEXT_CSV = "text/csv";
+	private static final String SYS_FILE_META_ID = "sys_FileMeta_ID";
+	public enum Permission
+	{
+		READ, WRITE, COUNT, NONE, WRITEMETA
+	}
+
+	// Request parameters1
 	private static final String PATH = "api/v1/";
-
-	// Admin credentioals
-	private static final String DEFAULT_HOST = "http://localhost:8080";
-	private static final String DEFAULT_ADMIN_NAME = "admin";
-	private static final String DEFAULT_ADMIN_PW = "admin";
-
 	// Permission identifiers
-	private static final String PACKAGE_PERMISSION_ID = "package_permission_ID";
-	private static final String ENTITY_TYPE_PERMISSION_ID = "entityType_permission_ID";
-	private static final String ATTRIBUTE_PERMISSION_ID = "attribute_permission_ID";
 	private static final String FREEMARKER_TEMPLATE_PERMISSION_ID = "freemarkerTemplate_permission_ID";
-	private static final String SCRIPT_TYPE_PERMISSION_ID = "scriptType_permission_ID";
-	private static final String TYPE_TEST_PERMISSION_ID = "typeTest_permission_ID";
-	private static final String TYPE_TEST_REF_PERMISSION_ID = "typeTestRef_permission_ID";
-	private static final String LOCATION_PERMISSION_ID = "location_permission_ID";
-	private static final String PERSONS_PERMISSION_ID = "persons_permission_ID";
 
+	private static final String SCRIPT_TYPE_PERMISSION_ID = "scriptType_permission_ID";
 	// User credentials
 	private static final String REST_TEST_USER = "rest_test_user";
 	private static final String REST_TEST_USER_PASSWORD = "rest_test_user_password";
@@ -87,24 +62,18 @@ public class RestControllerIT
 		String adminToken = login(adminUserName, adminPassword);
 
 		LOG.info("Importing RestControllerV1_TestEMX.xlsx...");
-		uploadEMX(adminToken);
+		uploadEMX(adminToken, "/RestControllerV1_TestEMX.xlsx");
 		LOG.info("Importing Done");
 
 		createUser(adminToken, REST_TEST_USER, REST_TEST_USER_PASSWORD);
 
-		String testUserId = getUserId(adminToken, REST_TEST_USER);
+		String testUserId = getUserId(adminToken, PATH, REST_TEST_USER);
 		LOG.info("testUserId: " + testUserId);
 
-		grantSystemRights(adminToken, FREEMARKER_TEMPLATE_PERMISSION_ID, testUserId, "sys_FreemarkerTemplate");
-		grantSystemRights(adminToken, SCRIPT_TYPE_PERMISSION_ID, testUserId, "sys_scr_ScriptType");
-		grantSystemRights(adminToken, PACKAGE_PERMISSION_ID, testUserId, "sys_md_Package");
-		grantSystemRights(adminToken, ENTITY_TYPE_PERMISSION_ID, testUserId, "sys_md_EntityType");
-		grantSystemRights(adminToken, ATTRIBUTE_PERMISSION_ID, testUserId, "sys_md_Attribute");
-
-		grantRights(adminToken, TYPE_TEST_PERMISSION_ID, testUserId, "TypeTest");
-		grantRights(adminToken, TYPE_TEST_REF_PERMISSION_ID, testUserId, "TypeTestRef");
-		grantRights(adminToken, LOCATION_PERMISSION_ID, testUserId, "Location");
-		grantRights(adminToken, PERSONS_PERMISSION_ID, testUserId, "Person");
+		grantSystemRights(adminToken, PATH, FREEMARKER_TEMPLATE_PERMISSION_ID, testUserId, "sys_FreemarkerTemplate", Permission.WRITE);
+		grantSystemRights(adminToken, PATH, SCRIPT_TYPE_PERMISSION_ID, testUserId, "sys_scr_ScriptType", Permission.READ);
+		grantSystemRights(adminToken, PATH, SYS_SEC_USER_AUTHORITY_ID, testUserId, "sys_sec_UserAuthority", Permission.COUNT);
+		grantSystemRights(adminToken, PATH, SYS_FILE_META_ID, testUserId, "sys_FileMeta", Permission.WRITEMETA);
 
 		// Add home plugin
 		// Add Language entity
@@ -112,353 +81,127 @@ public class RestControllerIT
 		this.testUserToken = login(REST_TEST_USER, REST_TEST_USER_PASSWORD);
 	}
 
-	/**
-	 * Import TypeTest, TypeTestRef, Location and Person
-	 * using add/update
-	 *
-	 * @param token to use for login
-	 */
-	private void uploadEMX(String token)
+	@Test
+	public void getWithoutTokenNotAllowed()
 	{
-		URL resourceUrl = Resources.getResource(RestControllerIT.class, "/RestControllerV1_TestEMX.xlsx");
-		File file = null;
-		try
-		{
-			file = new File(new URI(resourceUrl.toString()).getPath());
-		}
-		catch (URISyntaxException e)
-		{
-			LOG.error(e.getMessage());
-		}
+		ValidatableResponse response;
 
-		given().log().all().multiPart(file).param("file").param("action", "ADD_UPDATE_EXISTING")
-				.header(X_MOLGENIS_TOKEN, token).post("plugin/importwizard/importFile").then().log().all()
-				.statusCode(201);
-	}
+		response = getWithoutToken("sys_FreemarkerTemplate");
+		response.statusCode(UNAUTHORIZED).body("errors.message[0]", equalTo("No [COUNT] permission on entity [sys_FreemarkerTemplate]"));
 
-	/**
-	 * Grant user rights in list of entities
-	 *
-	 * @param adminToken the token to use for signin
-	 * @param userId     the ID (not the name) of the user that needs to get the rights
-	 * @param entity     a list of entity names
-	 * @return int http response code
-	 */
-	private int grantRights(String adminToken, String permissionID, String userId, String entity)
-	{
-		return grantSystemRights(adminToken, permissionID, userId,
-				getEntityTypeId(adminToken, "name", entity, "sys_md_EntityType"));
-	}
+		response = getWithoutToken("sys_scr_ScriptType");
+		response.statusCode(UNAUTHORIZED).body("errors.message[0]", equalTo("No [COUNT] permission on entity [sys_scr_ScriptType]"));
 
-	private int grantSystemRights(String adminToken, String permissionID, String userId, String entity)
-	{
-		String right = "ROLE_ENTITY_WRITE_" + entity;
-		JSONObject body = new JSONObject(ImmutableMap.of("id", permissionID, "role", right, "User", userId));
+		response = getWithoutToken("sys_sec_UserAuthority");
+		response.statusCode(UNAUTHORIZED).body("errors.message[0]", equalTo("No [COUNT] permission on entity [sys_sec_UserAuthority]"));
 
-		return given().log().all().header("x-molgenis-token", adminToken).contentType(APPLICATION_JSON)
-				.body(body.toJSONString()).when().post(PATH + "sys_sec_UserAuthority").then().log().all().extract()
-				.statusCode();
-	}
-
-	private String getUserId(String adminToken, String userName)
-	{
-		return getEntityTypeId(adminToken, "username", userName, "sys_sec_User");
-	}
-
-	private String getEntityTypeId(String adminToken, String attribute, String value, String entityName)
-	{
-		Map<String, Object> query = of("q",
-				singletonList(of("field", attribute, "operator", "EQUALS", "value", value)));
-		JSONObject body = new JSONObject(query);
-
-		return given().header("x-molgenis-token", adminToken).contentType(APPLICATION_JSON).queryParam("_method", "GET")
-				.body(body.toJSONString()).when().post(PATH + entityName).then().log().all().extract()
-				.path("items[0].id");
+		response = getWithoutToken("sys_sec_GroupAuthority");
+		response.statusCode(UNAUTHORIZED).body("errors.message[0]", equalTo("No [COUNT] permission on entity [sys_sec_GroupAuthority]"));
 	}
 
 	@Test
-	public void testFreemarkerNotAllowed()
+	public void getWithTokenIsAllowed()
 	{
-		noPermissionRequest("sys_FreemarkerTemplate");
+		getWithToken("sys_FreemarkerTemplate", this.testUserToken).log().body().statusCode(200);
+		getWithToken("sys_scr_ScriptType", this.testUserToken).log().body().statusCode(200);
 	}
 
 	@Test
-	public void testScriptTypeNotAllowed()
+	public void getWithoutReadPermissionIsNotAllowed()
 	{
-		noPermissionRequest("sys_scr_ScriptType");
+		ValidatableResponse response = getWithToken("sys_sec_UserAuthority", this.testUserToken);
+		response.statusCode(UNAUTHORIZED).body("errors.message[0]", equalTo("No [READ] permission on entity [sys_sec_UserAuthority]"));
 	}
 
 	@Test
-	public void testUserAuthorityNotAllowed()
+	public void getWithoutCountPermissionIsNotAllowed()
 	{
-		noPermissionRequest("sys_sec_UserAuthority");
+		ValidatableResponse response = getWithToken("sys_sec_GroupAuthority", this.testUserToken);
+		response.statusCode(UNAUTHORIZED).body("errors.message[0]", equalTo("No [COUNT] permission on entity [sys_sec_GroupAuthority]"));
 	}
 
 	@Test
-	public void testGroupAuthorityNotAllowed()
+	public void deleteNonExistingEntity()
 	{
-		noPermissionRequest("sys_sec_GroupAuthority");
+		given().log().uri()
+				.header(X_MOLGENIS_TOKEN, this.testUserToken)
+				.when().delete(PATH + "sys_FileMeta" + "/non-existing-entity_id")
+				.then().log().all()
+				.statusCode(NO_CONTENT);
 	}
 
 	@Test
-	public void testGetFreemarkerSuccess()
+	public void deleteWithoutWritePermissionFails ()
 	{
-		successRequest("sys_FreemarkerTemplate", this.testUserToken);
+		given().log().method().log().uri()
+				.header(X_MOLGENIS_TOKEN, this.testUserToken)
+				.when().delete(PATH + "sys_scr_ScriptType/R")
+				.then().statusCode(UNAUTHORIZED)
+				.body("errors.message[0]", equalTo("No [WRITE] permission on entity [sys_scr_ScriptType]"));
+
+		given().log().method().log().uri()
+				.header(X_MOLGENIS_TOKEN, this.testUserToken)
+				.when().delete(PATH + "sys_sec_UserAuthority")
+				.then().statusCode(UNAUTHORIZED)
+				.body("errors.message[0]", equalTo("No [WRITE] permission on entity [sys_sec_UserAuthority]"));
+
+		given().log().method().log().uri()
+				.header(X_MOLGENIS_TOKEN, this.testUserToken)
+				.when().delete(PATH + "sys_sec_GroupAuthority")
+				.then().statusCode(UNAUTHORIZED)
+				.body("errors.message[0]", equalTo("No [WRITE] permission on entity [sys_sec_GroupAuthority]"));
 	}
 
 	@Test
-	public void testGetScriptTypeSuccess()
+	public void logoutWithoutTokenFails ()
 	{
-		successRequest("sys_scr_ScriptType", this.testUserToken);
+		given()
+				.log().uri().log().method()
+				.when().post(PATH + "logout")
+				.then().statusCode(BAD_REQUEST)
+				.log().all()
+				.body("errors.message[0]", equalTo("Missing token in header"));
 	}
 
 	@Test
-	public void testEntityExists()
+	public void logoutWithToken()
 	{
-		given().log().all().header(X_MOLGENIS_TOKEN, this.testUserToken).contentType(TEXT_PLAIN).when()
-				.get(PATH + "it_emx_datatypes_TypeTestRef/exist").then().log().all().statusCode(200)
-				.body(equalTo("true"));
+		given().log().uri().log().method()
+				.header(X_MOLGENIS_TOKEN, this.testUserToken)
+				.when().post(PATH + "logout")
+				.then().statusCode(OKE)
+				.log().all()
+		;
+
+		given().log().uri().log().method()
+				.header(X_MOLGENIS_TOKEN, this.testUserToken)
+				.when().get(PATH + "sys_FreemarkerTemplate")
+				.then().statusCode(UNAUTHORIZED)
+				.body("errors.message[0]", equalTo("No [COUNT] permission on entity [sys_FreemarkerTemplate]"));
+
+		given().log().uri().log().method()
+				.when().get(PATH + "sys_FreemarkerTemplate")
+				.then().statusCode(UNAUTHORIZED)
+				.body("errors.message[0]", equalTo("No [COUNT] permission on entity [sys_FreemarkerTemplate]"));
 	}
 
-	@Test
-	public void testEntityNotExists()
+
+
+	private ValidatableResponse getWithoutToken(String requestedEntity)
 	{
-		given().log().all().header(X_MOLGENIS_TOKEN, this.testUserToken).contentType(TEXT_PLAIN).when()
-				.get(PATH + "sys_NonExistingEntity/exist").then().log().all().statusCode(200).body(equalTo("false"));
+		return given().log().uri().when().get(PATH + requestedEntity).then();
 	}
 
-	@Test
-	public void testGetEntityType()
+	private ValidatableResponse getWithToken(String requestedEntity, String token)
 	{
-		ValidatableResponse response = given().log().all().header(X_MOLGENIS_TOKEN, this.testUserToken)
-				.contentType(APPLICATION_JSON).when().get(PATH + "it_emx_datatypes_TypeTestRef/meta").then().log()
-				.all();
-		validateGetEntityType(response);
+		return given().log().all()
+				.header(X_MOLGENIS_TOKEN, token)
+				.contentType(APPLICATION_JSON).when()
+				.get(PATH + requestedEntity).then();
+
 	}
 
-	@Test
-	public void testGetEntityTypePost()
-	{
-		ValidatableResponse response = given().log().all().header(X_MOLGENIS_TOKEN, this.testUserToken)
-				.contentType(APPLICATION_JSON).body(new EntityTypeRequest()).when()
-				.post(PATH + "it_emx_datatypes_TypeTestRef/meta?_method=GET").then().log().all();
-		validateGetEntityType(response);
-	}
 
-	@Test
-	public void testRetrieveEntityAttributeMeta()
-	{
-		ValidatableResponse response = given().log().all().header(X_MOLGENIS_TOKEN, this.testUserToken)
-				.contentType(APPLICATION_JSON).when().get(PATH + "sys_scr_ScriptType/meta/name").then().log().all();
-		validateRetrieveEntityAttributeMeta(response);
-	}
-
-	@Test
-	public void testRetrieveEntityAttributeMetaPost()
-	{
-		ValidatableResponse response = given().log().all().header(X_MOLGENIS_TOKEN, this.testUserToken)
-				.contentType(APPLICATION_JSON).body(new EntityTypeRequest()).when()
-				.post(PATH + "sys_scr_ScriptType/meta/name?_method=GET").then().log().all();
-		validateRetrieveEntityAttributeMeta(response);
-	}
-
-	@Test
-	public void testRetrieveEntity()
-	{
-		ValidatableResponse response = given().log().all().header(X_MOLGENIS_TOKEN, this.testUserToken)
-				.contentType(APPLICATION_JSON).when().get(PATH + "it_emx_datatypes_TypeTestRef/ref1").then().log()
-				.all();
-		validateRetrieveEntity(response);
-	}
-
-	@Test
-	public void testRetrieveEntityPost()
-	{
-		ValidatableResponse response = given().log().all().header(X_MOLGENIS_TOKEN, this.testUserToken)
-				.contentType(APPLICATION_JSON).body(new EntityTypeRequest()).when()
-				.post(PATH + "it_emx_datatypes_TypeTestRef/ref1?_method=GET").then().log().all();
-		validateRetrieveEntity(response);
-	}
-
-	@Test
-	public void testRetrieveEntityAttribute()
-	{
-		ValidatableResponse response = given().log().all().header(X_MOLGENIS_TOKEN, this.testUserToken)
-				.contentType(APPLICATION_JSON).when().get(PATH + "it_emx_datatypes_TypeTest/1/xxref_value").then().log()
-				.all();
-		validateRetrieveEntityAttribute(response);
-	}
-
-	@Test
-	public void testRetrieveEntityAttributePost()
-	{
-		ValidatableResponse response = given().log().all().header(X_MOLGENIS_TOKEN, this.testUserToken)
-				.contentType(APPLICATION_JSON).body(new EntityCollectionRequest()).when()
-				.post(PATH + "it_emx_datatypes_TypeTest/1/xxref_value?_method=GET").then().log().all();
-		validateRetrieveEntityAttribute(response);
-	}
-
-	@Test
-	public void testRetrieveEntityCollectionResponse()
-	{
-		ValidatableResponse response = given().log().all().header(X_MOLGENIS_TOKEN, this.testUserToken)
-				.contentType(APPLICATION_JSON).when().get(PATH + "it_emx_datatypes_TypeTestRef").then().log().all();
-		validateRetrieveEntityCollectionResponse(response);
-	}
-
-	@Test
-	public void testRetrieveEntityCollectionResponsePost()
-	{
-		ValidatableResponse response = given().log().all().header(X_MOLGENIS_TOKEN, this.testUserToken)
-				.contentType(APPLICATION_JSON).body(new EntityCollectionRequest()).when()
-				.post(PATH + "it_emx_datatypes_TypeTestRef?_method=GET").then().log().all();
-		validateRetrieveEntityCollectionResponse(response);
-	}
-
-	@Test(enabled = false)
-	// FIXME valdate output in CSV
-	public void testRetrieveEntityCollection()
-	{
-		given().log().all().header(X_MOLGENIS_TOKEN, this.testUserToken).contentType(TEXT_CSV).when()
-				.get(PATH + "csv/sys_scr_ScriptType").then().contentType(TEXT_CSV).log().all().statusCode(200)
-				.body(equalTo(""));
-	}
-
-	@Test
-	public void testCreateFromFormPost()
-	{
-		// Add new entity from form post
-		given().log().all().header(X_MOLGENIS_TOKEN, this.testUserToken).contentType(APPLICATION_FORM_URL_ENCODED)
-				.formParam("value", "ref6").formParam("label", "label6").when()
-				.post(PATH + "it_emx_datatypes_TypeTestRef").then().log().all().statusCode(201);
-
-		// Check if entity was added
-		given().log().all().header(X_MOLGENIS_TOKEN, this.testUserToken).contentType(APPLICATION_JSON).when()
-				.get(PATH + "it_emx_datatypes_TypeTestRef/ref6").then().log().all().statusCode(200)
-				.body("href", equalTo("/api/v1/it_emx_datatypes_TypeTestRef/ref6"), "value", equalTo("ref6"), "label",
-						equalTo("label6"));
-	}
-
-	@Test(enabled = false)
-	// FIXME 500 error
-	public void testCreateFromFormPostMultiPart()
-	{
-		// Add new entity from multipart form post
-		given().log().all().config(config()
-				.encoderConfig(encoderConfig().encodeContentTypeAs("multipart/form-data", ContentType.JSON)))
-				.header(X_MOLGENIS_TOKEN, this.testUserToken).contentType("multipart/form-data")
-				.formParam("name", "IT_ScriptType").when().post(PATH + "sys_scr_ScriptType").then().log().all()
-				.statusCode(201);
-
-		// Check if entity was added
-		String responseBody = "{\"href\":\"/api/v1/sys_scr_ScriptType/IT_ScriptType\",\"name\":\"IT_ScriptType\"}";
-		given().log().all().header(X_MOLGENIS_TOKEN, this.testUserToken).contentType(APPLICATION_JSON).when()
-				.get(PATH + "sys_scr_ScriptType/IT_ScriptType").then().log().all().statusCode(200)
-				.body(equalTo(responseBody));
-
-		// Remove entity
-		given().log().all().header(X_MOLGENIS_TOKEN, this.testUserToken)
-				.delete("api/v2/sys_scr_ScriptType/IT_ScriptType").then().log().all().statusCode(204);
-	}
-
-	@Test(enabled = false)
-	// FIXME com.fasterxml.jackson.databind.JsonMappingException: No serializer found for class java.io.ByteArrayOutputStream and no properties discovered to create BeanSerializer (to avoid exception, disable SerializationFeature.FAIL_ON_EMPTY_BEANS) ) (through reference chain: org.springframework.mock.web.MockHttpServletResponse["outputStream"]->org.springframework.mock.web.ResponseServletOutputStream["targetStream"])
-	public void testCreate()
-	{
-		Map<String, Object> entityMap = newHashMap();
-		entityMap.put("name", "IT_ScriptType");
-
-		// Add new entity from multipart form post
-		given().log().all().header(X_MOLGENIS_TOKEN, this.testUserToken).body(entityMap)
-				.formParam("response", new MockHttpServletResponse()).when().post(PATH + "sys_scr_ScriptType").then()
-				.log().all().statusCode(201);
-
-		// Check if entity was added
-		String responseBody = "{\"href\":\"/api/v1/sys_scr_ScriptType/IT_ScriptType\",\"name\":\"IT_ScriptType\"}";
-		given().log().all().header(X_MOLGENIS_TOKEN, this.testUserToken).contentType(APPLICATION_JSON).when()
-				.get(PATH + "sys_scr_ScriptType/IT_ScriptType").then().log().all().statusCode(200)
-				.body(equalTo(responseBody));
-
-		// Remove entity
-		given().log().all().header(X_MOLGENIS_TOKEN, this.testUserToken)
-				.delete("api/v2/sys_scr_ScriptType/IT_ScriptType").then().log().all().statusCode(204);
-	}
-
-	private void noPermissionRequest(String requestedEntity)
-	{
-		given().log().all().when().get(PATH + requestedEntity).then().statusCode(401)
-				.body("errors.message[0]", equalTo("No [COUNT] permission on entity [" + requestedEntity + "]"));
-	}
-
-	private void successRequest(String requestedEntity, String token)
-	{
-		given().log().all().header(X_MOLGENIS_TOKEN, token).contentType(APPLICATION_JSON).when()
-				.get(PATH + requestedEntity).then().log().all().statusCode(200);
-	}
-
-	private void validateGetEntityType(ValidatableResponse response)
-	{
-		response.statusCode(200);
-		response.body("href", equalTo("/api/v1/it_emx_datatypes_TypeTestRef/meta"), "hrefCollection",
-				equalTo("/api/v1/it_emx_datatypes_TypeTestRef"), "name", equalTo("it_emx_datatypes_TypeTestRef"),
-				"label", equalTo("TypeTestRef"), "description", equalTo("MOLGENIS Data types test ref entity"),
-				"attributes.value.href", equalTo("/api/v1/it_emx_datatypes_TypeTestRef/meta/value"),
-				"attributes.label.href", equalTo("/api/v1/it_emx_datatypes_TypeTestRef/meta/label"), "labelAttribute",
-				equalTo("label"), "idAttribute", equalTo("value"), "lookupAttributes",
-				equalTo(newArrayList("value", "label")), "isAbstract", equalTo(false), "languageCode", equalTo("en"),
-				"writable", equalTo(true));
-	}
-
-	private void validateRetrieveEntityAttributeMeta(ValidatableResponse response)
-	{
-		response.statusCode(200);
-		response.body("href", equalTo("/api/v1/sys_scr_ScriptType/meta/name"), "fieldType", equalTo("STRING"), "name",
-				equalTo("name"), "label", equalTo("name"), "attributes", equalTo(newArrayList()), "enumOptions",
-				equalTo(newArrayList()), "maxLength", equalTo(255), "auto", equalTo(false), "nillable", equalTo(false),
-				"readOnly", equalTo(true), "labelAttribute", equalTo(true), "unique", equalTo(true), "visible",
-				equalTo(true), "lookupAttribute", equalTo(false), "isAggregatable", equalTo(false));
-	}
-
-	private void validateRetrieveEntity(ValidatableResponse response)
-	{
-		response.statusCode(200);
-		response.body("href", equalTo("/api/v1/it_emx_datatypes_TypeTestRef/ref1"), "value", equalTo("ref1"), "label",
-				equalTo("label1"));
-	}
-
-	private void validateRetrieveEntityAttribute(ValidatableResponse response)
-	{
-		response.statusCode(200);
-		response.body("href", equalTo("/api/v1/it_emx_datatypes_TypeTestRef/1/xxref_value"), "value", equalTo("ref1"),
-				"label", equalTo("label1"));
-	}
-
-	private void validateRetrieveEntityCollectionResponse(ValidatableResponse response)
-	{
-		response.statusCode(200);
-		response.body("href", equalTo("/api/v1/it_emx_datatypes_TypeTestRef"), "meta.href",
-				equalTo("/api/v1/it_emx_datatypes_TypeTestRef/meta"), "meta.hrefCollection",
-				equalTo("/api/v1/it_emx_datatypes_TypeTestRef"), "meta.name", equalTo("it_emx_datatypes_TypeTestRef"),
-				"meta.label", equalTo("TypeTestRef"), "meta.description",
-				equalTo("MOLGENIS Data types test ref entity"), "meta.attributes.value.href",
-				equalTo("/api/v1/it_emx_datatypes_TypeTestRef/meta/value"), "meta.attributes.label.href",
-				equalTo("/api/v1/it_emx_datatypes_TypeTestRef/meta/label"), "meta.labelAttribute", equalTo("label"),
-				"meta.idAttribute", equalTo("value"), "meta.lookupAttributes", equalTo(newArrayList("value", "label")),
-				"meta.isAbstract", equalTo(false), "meta.languageCode", equalTo("en"), "meta.writable", equalTo(true),
-				"start", equalTo(0), "num", equalTo(100), "total", equalTo(6), "items[0].href",
-				equalTo("/api/v1/it_emx_datatypes_TypeTestRef/ref6"), "items[0].value", equalTo("ref6"),
-				"items[0].label", equalTo("label6"), "items[1].href",
-				equalTo("/api/v1/it_emx_datatypes_TypeTestRef/ref1"), "items[1].value", equalTo("ref1"),
-				"items[1].label", equalTo("label1"), "items[2].href",
-				equalTo("/api/v1/it_emx_datatypes_TypeTestRef/ref2"), "items[2].value", equalTo("ref2"),
-				"items[2].label", equalTo("label2"), "items[3].href",
-				equalTo("/api/v1/it_emx_datatypes_TypeTestRef/ref3"), "items[3].value", equalTo("ref3"),
-				"items[3].label", equalTo("label3"), "items[4].href",
-				equalTo("/api/v1/it_emx_datatypes_TypeTestRef/ref4"), "items[4].value", equalTo("ref4"),
-				"items[4].label", equalTo("label4"), "items[5].href",
-				equalTo("/api/v1/it_emx_datatypes_TypeTestRef/ref5"), "items[5].value", equalTo("ref5"),
-				"items[5].label", equalTo("label5"));
-	}
 
 	@AfterClass
 	public void afterClass()
